solution1/backend/requirements.txt
fastapi
azurefunctions-extensions-http-fastapi
starlette
azure-functions
pyodbc
openai
uvicorn
starlette
openai
msal
requests
langchain
langgraph
pyjwt
azure.identity
azure.keyvault.secrets
pydantic
typing-extensions
opencensus-ext-azure
opencensus
sqlalchemy
